+++ docs/source/changes.md
@@ -9,6 +9,8 @@ releases are available on [PyPI](https://pypi.org/project/pytask) and
 
 - {pull}`345` updates the version numbers in animations.
 - {pull}`352` publishes `db` that is required by pytask-environment.
+- {pull}`354` adds a `-f/--force` flag to execute tasks even though nothing may have
+  changed.
 - {pull}`355` refactors a lot of things related to nodes.
 
 ## 0.3.1 - 2023-12-25